# Conflicts:
#	sandbox-backend/src/main/java/com/exadel/team2/sandbox/web/employee/CreateEmployeeDto.java
#	sandbox-backend/src/main/java/com/exadel/team2/sandbox/web/employee/UpdateEmployeeDto.java
#	sandbox-backend/src/main/resources/db/changelog/db.changelog-master.yaml
#	sandbox-web/build.gradle

+++ sandbox-backend/src/main/java/com/exadel/team2/sandbox/service/impl/CandidateEventServiceImpl.java
@@ -8,6 +8,7 @@ import com.exadel.team2.sandbox.dao.rsql.CustomRsqlVisitor;
 import com.exadel.team2.sandbox.entity.CandidateEntity;
 import com.exadel.team2.sandbox.entity.CandidateEventEntity;
 import com.exadel.team2.sandbox.entity.EventEntity;
+import com.exadel.team2.sandbox.exceptions.NoSuchException;
 import com.exadel.team2.sandbox.mapper.CandidateEventMapper;
 import com.exadel.team2.sandbox.service.CandidateEventService;
 import com.exadel.team2.sandbox.web.candidate_event.CreateCandidateEventDto;
@@ -15,7 +16,6 @@ import com.exadel.team2.sandbox.web.candidate_event.ResponseCandidateEventDto;
 import cz.jirutka.rsql.parser.RSQLParser;
 import cz.jirutka.rsql.parser.ast.Node;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -25,10 +25,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
-import java.util.List;
-import java.util.stream.Collectors;
 
-@Slf4j
 @Transactional
 @Service
 @RequiredArgsConstructor
@@ -46,36 +43,11 @@ public class CandidateEventServiceImpl implements CandidateEventService {
     }
 
     @Override
-    public List<ResponseCandidateEventDto> getAll() {
-        List<CandidateEventEntity> candidateEventEntities = candidateEventDAO.findAll();
-        if (candidateEventEntities.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NO_CONTENT, "No Content");
-        }
-        return candidateEventEntities.stream()
-                .map(candidateEventMapper::convertEntityToDto)
-                .collect(Collectors.toList());
-    }
-
-    @Override
-    public ResponseCandidateEventDto save(CreateCandidateEventDto createCandidateEventDto) {
-        CandidateEventEntity candidateEventEntity = candidateEventMapper.convertDtoToEntity(createCandidateEventDto);
-
-
-        EventEntity eventEntity = eventDAO.findById(createCandidateEventDto.getIdEvent())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
-        candidateEventEntity.setEvent(eventEntity);
-
-        CandidateEntity candidateEntity = candidateDAO.findById(createCandidateEventDto.getIdCandidate())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Candidate not found"));
-        candidateEventEntity.setCandidates((List<CandidateEntity>) candidateEntity);
-
-        candidateEventEntity.setCreatedAt(LocalDateTime.now());
-        candidateEventDAO.save(candidateEventEntity);
-        return candidateEventMapper.convertEntityToDto(candidateEventEntity);
-    }
-
-    @Override
     public void delete(Long id) {
+        if (!candidateEventDAO.existsById(id)) {
+            throw new NoSuchException("CandidateEvent with ID = " + id + " not found in Database. " +
+                    "Unable to delete an CandidateEvent that does not exist.");
+        }
         candidateEventDAO.deleteById(id);
     }
 
@@ -89,5 +61,18 @@ public class CandidateEventServiceImpl implements CandidateEventService {
             return candidateEventDAO.findAll(spec, pageable)
                     .map(candidateEventMapper::convertEntityToDto);
         }
+    }
+
+    @Override
+    public ResponseCandidateEventDto save(CreateCandidateEventDto createCandidateEventDto) {
+        CandidateEventEntity candidateEventEntity = candidateEventMapper.convertDtoToEntity(createCandidateEventDto);
+        EventEntity eventEntity = eventDAO.findById(createCandidateEventDto.getIdEvent())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Event not found"));
+        candidateEventEntity.setEvent(eventEntity);
+        CandidateEntity candidateEntity = candidateDAO.findById(createCandidateEventDto.getIdCandidate())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Candidate not found"));
+        candidateEventEntity.setCandidate(candidateEntity);
+        candidateEventDAO.save(candidateEventEntity);
+        return candidateEventMapper.convertEntityToDto(candidateEventEntity);
     }
 }
