+++ sandbox-backend/src/main/java/com/exadel/team2/sandbox/service/impl/InterviewFeedbackServiceImpl.java
@@ -8,6 +8,7 @@ import com.exadel.team2.sandbox.dao.rsql.CustomRsqlVisitor;
 import com.exadel.team2.sandbox.entity.CandidateEntity;
 import com.exadel.team2.sandbox.entity.EmployeeEntity;
 import com.exadel.team2.sandbox.entity.InterviewFeedbackEntity;
+import com.exadel.team2.sandbox.exceptions.NoSuchException;
 import com.exadel.team2.sandbox.mapper.InterviewFeedbackMapper;
 import com.exadel.team2.sandbox.service.InterviewFeedbackService;
 import com.exadel.team2.sandbox.web.interview_feedback.CreateInterviewFeedbackDto;
@@ -16,7 +17,6 @@ import com.exadel.team2.sandbox.web.interview_feedback.UpdateInterviewFeedbackDt
 import cz.jirutka.rsql.parser.RSQLParser;
 import cz.jirutka.rsql.parser.ast.Node;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -25,11 +25,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Slf4j
 @Transactional
 @Service
 @RequiredArgsConstructor
@@ -60,47 +58,38 @@ public class InterviewFeedbackServiceImpl implements InterviewFeedbackService {
 
     @Override
     public ResponseInterviewFeedbackDto save(CreateInterviewFeedbackDto createInterviewFeedbackDTO) {
-
         InterviewFeedbackEntity interviewFeedbackEntity = interviewFeedbackMapper.convertDtoToEntity(createInterviewFeedbackDTO);
-
         EmployeeEntity employeeEntity = employeeDAO.findById(createInterviewFeedbackDTO.getIdEmployee())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee not found"));
         interviewFeedbackEntity.setEmployee(employeeEntity);
-
         CandidateEntity candidateEntity = candidateDAO.findById(createInterviewFeedbackDTO.getIdCandidate())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Candidate not found"));
         interviewFeedbackEntity.setCandidate(candidateEntity);
-
         interviewFeedbackEntity.setFeedback(createInterviewFeedbackDTO.getFeedback());
-        interviewFeedbackEntity.setDateOfCreate(LocalDateTime.now());
-        interviewFeedbackEntity.setDateOfUpdate(LocalDateTime.now());
         interviewFeedbackDAO.save(interviewFeedbackEntity);
-
         return interviewFeedbackMapper.convertEntityToDto(interviewFeedbackEntity);
     }
 
     @Override
-    public ResponseInterviewFeedbackDto update(long id, UpdateInterviewFeedbackDto updateInterviewFeedbackDto) {
+    public ResponseInterviewFeedbackDto update(Long id, UpdateInterviewFeedbackDto updateInterviewFeedbackDto) {
         InterviewFeedbackEntity interviewFeedbackEntity = interviewFeedbackDAO.findById(id)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "InterviewFeedback not found"));
-
         EmployeeEntity employeeEntity = employeeDAO.findById(updateInterviewFeedbackDto.getIdEmployee())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Employee not found"));
         interviewFeedbackEntity.setEmployee(employeeEntity);
-
         CandidateEntity candidateEntity = candidateDAO.findById(updateInterviewFeedbackDto.getIdCandidate())
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Candidate not found"));
         interviewFeedbackEntity.setCandidate(candidateEntity);
-
-
         interviewFeedbackEntity.setFeedback(updateInterviewFeedbackDto.getFeedback());
-        interviewFeedbackEntity.setDateOfUpdate(LocalDateTime.now());
-
         return interviewFeedbackMapper.convertEntityToDto(interviewFeedbackDAO.save(interviewFeedbackEntity));
     }
 
     @Override
     public void delete(Long id) {
+        if (!interviewFeedbackDAO.existsById(id)) {
+            throw new NoSuchException("interviewFeedback with ID = " + id + " not found in Database. " +
+                    "Unable to delete an feedback that does not exist.");
+        }
         interviewFeedbackDAO.deleteById(id);
     }
